BUG FIX: Translating third-body colliders' names
When reading in third-body colliders from a pdepreactions.txt file, RMG
would not recognize these species as the core/edge species within the
reaction mechanism (instead, these names were treated as strings).  This
inconsistency would be translated to the chem.inp file, causing a problem
when run through the CHEMKIN pre-processor.

Now, the strings are translated into the respective species within the
model, including adding the chemkinName to the chem.inp file.

Note: There are two assumptions in this new step
   (1) The third-body colliders chemgraphs are defined in the species.txt file
   (2) The strings "ar", "ne", "he", and "n2" are safe and recognized as inert
   gases

If both of these fail, RMG will print out the name of the unrecognized species
and will terminate the simulation.

Closes #86