Use StringBuffer for string concatenation when reading FAME output.
Strings are immutable. Adding a line onto the end of a string
is inefficient, and doing it many many thousands of times is a particularly
bad idea. Use a StringBuffer instead.
See http://www.yoda.arachsys.com/java/strings.html

Hopefully this closes #85 in which a fame result took about a day 
to be read in to a string.