Construct FAME input using StringBuilder.
StringBuilder should be more efficient than the string concatenation.

Also replaced Fame input reader with StringBuilder instead of StringBuffer
as the former is supposed to be faster in non-threaded situations
as it is not synchronized.  Also set the initial size to be a power of 2 
as this is reported to speed things up.