BUG FIX: PLOG Ea units should match with keyword following REACTIONS
In a previous commit

http://github.com/GreenGroup/RMG-Java/commit/26fb6bb8f4b9727eb6a20d7b81a1e9ca01eb12a8

MRH assumed that the PLOG Ea units were forced to be cal/mol.  Well, we all
know what happens when one assumes, and this case is no different.

As RHW pointed out, the keyword following the REACTIONS keyword in the
chem.inp file dictates what the Ea units for all reaction rate
coefficients should be, despite what the CHEMKIN-v.4.1.1 manual states.