BUG FIX in Sulfur database: 1,2_shiftS
Caleb was running a RMG simulation with diethysulfide.  When the simulation
tried to estimate the kinetics for the following reaction, generated by the
1,2_shiftS reaction family template, a RateConstantNotFound exception was
thrown:

HS-S-*CH2 --> *S-CH2-SH

The problem was in the definitions of the functional groups (dictionary.txt).
This family has three trees that it must find a match for.  One of the head
nodes is XSYJ, whose definition was:

1 *1 R!H 0 {2,S}
2 *2 S   0 {1,S} {3,S}
3 *3 R!H 1 {2,S}

However, another head node is X-Ss, whose definition is:

1 *1 C 0

In the above reaction, the reactant was labeled *1-*2-*3, in accord with the
XSYJ definition.  However, RMG could not find a match for X-Ss as this
definition forces the *1 atom to be a C (and RMG had already labeled a S
atom as *1).

MRH has changed the definition of XSYJ to:

1 *1 C   0 {2,S}
2 *2 S   0 {1,S} {3,S}
3 *3 R!H 1 {2,S}

This commit resolves the RateConstantNotFound exception encountered when
running the diethylsulfide input file.