Fixed bug in evaluating leak flux for small PDepNetworks.
For PDepNetworks of the form A + B -> C with C not included, we use the high
pressure limit to evaluate the leak flux. However, this code was written
assuming that C was always nonincluded, which is not the case for small
molecules like CH4 and H2O. This commit corrects this bug.