Updated rate coefficient for ROH + .OO. => HOO. + RO.
The RMG estimate for this type of reaction is poor (kinetics are much too
fast).  For the time being, Prof. Green has suggested the rate coefficient
expression that now exists in the database.  This estimate is certainly
more reasonable than the previous value estimated by RMG.

JDM and/or MRH will refine this value (see Issue #88).