Now using faster k(T,P) value as forward direction from FAME output.
This is consistent with RMG's convention of always using the
exothermic direction. Numerically it's better to divide a big number by
a big number to get a small number than to divide a small number by a
small number to get a big number. This is particularly important with
k(T,P) values, as they can be very small at very high or very low P.