Returning exothermic H_Abstraction reactions in chem.inp
To help avoid numerical issues when running CHEMKIN, RMG should always report
reactions in the exothermic direction.

This was true for most H_Abstraction reactions, unless one of the species
had a radical count >= 2.  In that case, this species would be labeled a
reactant.  Now, RMG always reports the exothermic direction; this will be
especially important with X-H + .O-O. --> HO2 + .X

The hard-coding for intra_H_migration has also been removed.  Before, RMG
would report the exothermic direction, unless the reaction was of the form
ROO. --> QOOH, in which case it would be reported as such.  Although this
family will be taken care of by JWA's recent addition, I am including it
in the event someone runs RMG w/pressure dependence turned off.