Adding boolean argument to readMeaningfulLine
Calling readMeaningfulLine automatically "trims" the line being read.
In some cases, having the leading/trailing white spaces is convenient.
For example, when reading in NASA polynomials, the coefficients are
blocked off in 15 character increments, including white space.

This looks like a rather large commit, but the majority of the changes have
been adding ", true" to all of the readMeaningfulLine calls throughout the
source code.