dasslAUTO saves species concentration profiles to SpeciesProfiles.txt
JDASSL.java wipes this file and writes the header line with the species names.
call_dasslAUTO.f90 appends to this file at every timestep.

The values are tab separated, and are in seconds (time) and moles per cm3 (species concentrations). The file should open directly in excel.
With auto time stepping the ODE solver determines the time steps and it saves
the values at each step.