FAME now accepts path reactions as always in the forward direction.
Previously, FAME worked best when the path reactions were provided in the
form A -> B or A -> B + C. This would be fine except that the reverse
kinetics fitting algorithm in the Java code is quite crude, and potentially
a source of error. Thus, FAME can now happily accept the kinetics in the
direction B + C -> A. For this case it will fit the reverse kinetics itself
- using linear least-squares regression - and use it to get k(E).

I've already seen at least one instance where the kinetics at 2000 K was
off by almost two orders of magnitude between the rough and fitted reverse
kinetics.

Closes #98.