Restoring the reporting of node names in chem.inp file
For each reaction rate coefficient expression in the chem.inp file, RMG again
reports the set of nodes found when traversing the tree.

RMG no longer reports the Keq.  MRH has added a "near" fix (see the commented
out code in this commit), however, the Keq expression would be reported ~6x
in the chem.inp file for each reaction.  This was due to the method
calculateTotalRate being called ~6x in an RMG run, for a given reaction.

NOTE (something to think about for the future, MRH has not thought about
it that much): I'm OK w/our needing to call this function multiple times.
However, why not store this value with each reaction.  Then, the next time
we call the method, first check to see if the value is not zero.

Closes #105