Bug fix: (Stop) Pruning species present in reaction/kinetics libraries
Before removing a pruned species from the SpeciesDictionary, check if it is
present in a Primary Kinetics / Reaction Library.  If so, do not remove it and
print a message to the user informing them which species will not be removed
from the SpeciesDictionary.

All other instances of the species are still removed.

Closes #107