Pruning non-kinetics/reaction library species & reactions
When pruning species and reactions from the edge of the reaction model, we
need to be careful not to delete:

(1) The graphs of any species present in a primary kinetics / reaction library
(2) The structures of any reactions present in a primary kinetics / reaction
    library

To check that we are not deleting "library" species, we pass a list of
species to pruneReactionModel().  To check for "library" reactions, MRH
has added code which checks the forward reaction's kinetics source string.
If the string contains the expression "Library", the reaction is either from
a kinetic or reaction library and its structure must remain in memory.  All
other instances of the reaction are removed from the reaction model (as
before).

Furthermore, there was a bug in the "fix" for Issue 107.  In coding the fix,
MRH assumed a primary kinetics library and reaction library were present
and hence did not check for null instances of either of these.  This commit
resolves this issue.

Closes #108