Bug fix: Additional don't prune kinetics / reaction library reactions
There was a typo in the previous commit, where the source of the forward
reaction was scanned for "Library" and the reverse was scanned for
"ReactionLibrary".  Both should be "Library".

Furthermore, this checks needs to be performed for the pdep portion of the
pruning code.  This commit resolves this issue.