Debugging pruning w/pressure-dependence
Stijn's recent butylformate job failed because of "Out of Memory".  Looking at
the log file, the last set of printed information stated there were ~42,000
species in the edge.  The pruning lines were present in the log file, but based
on the information present, not more than 500 were pruned from the model.  The
simulation then tried to expand the model, but failed shortly thereafter.

I should also note, Stijn asked for the edge to be reduced to 30,000 species.

After looking at the code, MRH thinks we may be labeling enough species to be
pruned from the model, however, after checking this list against the included
isomers in the pdepnetworks, this number is drastically reduced.

This commit prints out some statistics during the RMG pruning, to help
determine what (if anything) is amiss.