Removing "reaction-less" species from Dooley/C1 library
I've removed 5 species from the species.txt file in the Dooley/C1 kinetic
library.  None of these 5 are present in any reaction or pressure-dependent
reaction.

Note: If using this library as a Seed Mechanism, these species could
potentially have (identically) zero concentration and flux throughout a
simulation.  This may also be problematic when running with pruning, as
many species are labeled as unprunable.