Mike's code to compute (and print) kf and kr for all rxns in chem.inp file
This code reads in a chem.inp file (it assumes the thermo is present in the
same file, e.g. how RMG makes chem.inp files) and computes the kf and kr
for all pressure-independent reactions and pressure-dependent reactions (if
they are represented using Chebyshev polynomials).

The user needs to specify the temperature(s) and pressure(s) of interest.  The
default is 1 atm and 300-3000K (equally spaced on an inverse temperature
scale).

WARNING: May contain bugs.  If so, please let me know.