Safer reading of prunableSpecies list from ODE solver output.
I implemented a "1 or 0 else Error" rule rather than ">0 is true" rule.
There wasn't a bug here, but this should be safer.
Apologies for the System.exit(0) but there are no exception
classes for me to throw. Apologies also for not reducing the 
code duplication.

Also moved around some comments and indented some lines.