FAME now only returns k(T,P) in the faster direction.
This change was made in response to a situation where in one direction the
k(T,P) value was so low at one temperature that it went beyond the range of a
double-precision number and was therefore set to zero. The faster direction
was still slow, but large enough in magnitude to exist as a nonzero quantity.
This is important because the interpolation models work on a log k(T,P)
basis.

The validity check for k(T,P) values is now only performed for the direction
being written to the output file.

The faster direction is returned for numerical reasons; it is better to
divide a big number by a big number to get a small number than to divide a
small number by a small number to get a big number.