Some modifications due to no longer storing reverse k(T,P) values.
When the reverse PDepReaction is generated, it was bizarrely setting its
kinetics to be the same as the forward direction kinetics. Instead it now
sets them to null.

The calculateReverseFlux() method needed to be altered to use the forward
kinetics and equilibrium constant, since it cannot depend on having explicit
reverse kinetics.

The restart file now only writes PDepReaction objects in the forward
direction, since there are no kinetics to write in the reverse direction.

PDepReaction.calculateRate() now can get the rate in the reverse direction
using the forward direction rate and the equilibrium coefficient.