Catch to hopefully avoid k(T,P) values of 0 in FAME output.
For very large reaction networks it may come to pass that a k(T,P) value in
the forward direction is so low that it underflows a double-precision value
and is therefore set to zero. We can't fit any interpolation models in this
case since they use a log k(T,P) basis. However, these rates are so small
that they are negligible in the model. Thus, we simply set their values to a
very small (but not quite zero) value so we can still fit the interpolation
model.