Rewriting to use species leak fluxes instead of network leak fluxes.
Previously RMG would evaluate leak fluxes on a per-network basis, combining
all nonincluded species in that network. When enlarging, RMG would choose the
network with the highest leak flux and make the species in that network that
contributes the most included by exploring its reactions. This would cause
all partial networks containing that species as a nonincluded isomer to be
merged.

Instead, we want to evaluate leak fluxes on a per-species basis, combining
the leak fluxes from all networks that species appears in as a nonincluded
isomer. If the leak flux is the highest flux in the system, then that species
is made included (in all partial networks) and its path reactions explored.
This still causes all of the partial networks to be merged. The advantages of
this new approach are:

* Thinking in terms of species fluxes at all times rather than network
  fluxes.

* More consistency with pruning, which only uses species fluxes.

Significant refactoring of the PDep reaction model enlarger was required
because we can no longer distinguish between adding a species to the core and
enlarging a network by the type of the object in the update list.