Refactored RateBasedPDepRME.enlargeReactionModel() again.
We need to make the decisions about which species to add to the core or
explore in pressure-dependent networks for all reaction systems before we do
any enlarging, as otherwise there will be new edge species that will have
larger IDs than those that we have edge fluxes for. I had not accounted for
this when writing 125a79b6.

Instead we will maintain separate lists of species to add to core and species
to explore in networks. (We need two lists because we can't use the object
type any more to know whether we're adding to core or exploring in networks,
as everything is a Species object.)

Closes issue #117.