Using BufferedWriter instead of StringBuilder to write chem.inp
Stijn's recent job failed because of "Out of memory".  The error message from
java pointed to the toString() function within Chemkin.java.  In this
instance, we were trying to convert a StringBuilder to a String.  The
string happened to be the "REACTIONS" portion of a chem.inp file, which
contained >6,000 reactions (many of which were probably pdep).

I've changed the StringBuilder to a BufferedWriter for writing both the
chem.inp file and the tableOfRateCoefficients.txt file