Polyradical H-abstraction put in exothermic direction.
Previously there was some inconsistencies about when a structure has
two radicals or more on it, it was forced to remain a reactant (if it
was already one) else it was chosen reactant/product to make it exothermic.

Now, if both reactants and products have 2 or more radicals, we just do 
the exothermic rule. If only one of them has a polyradical, it is made
a reactant.

What to do here really depends on which parts of the database are best,
so it is hard to know what is the "correct" algorithm.

Also tidied indentation etc.