Better error message for forbidden species
When a species from the input file, or the RMG_database thermochemistry
folders, is marked as forbidden, a better error message is now supplied
to the user.

This error message only concerns read-in species that exceed the maximum
number of oxygens, radicals, or cycles (another already-in-place code checks
for the other options given to the user in the input file, e.g. maximum number
of carbons).