Bug fix: Converting InChIs (or .mol files) to adjacency lists
If the number of bonds in a molecule exceeded 99, the InChI2AdjList (or
Mol2AdjList) function would fail.  This was due to the number of atoms and
the number of bonds running together in the .mol file (as pointed out by
GRM).  As pointed out by RHW in the comments, this can be resolved by
recognizing every token takes 3 characters.

MRH has implemented and tested the fix.  The case presented by GRM in the
issue runs w/o error.

Closes #93