Remove --fpe-trap=overflow from Makefiles.
This causes FAME to throw a floating point exception, and yet after
trying to debug it for ages Josh an I gave up. If we don't compile with 
this flag, then it at least runs. There *may* be a bug in fame, but there
may also be a bug in the compiler. This should probably be looked at again
some time.