Remove --fpe-trap=zero,invalid from Makefiles (specifically for fame)
This is removed from the master makefile (and hence all called by it)
and fame's Makefile. It seems to only be a problem for fame.
It'd be nicer to find the cause of the floating point exception rather 
than just to not check for it, but for now this allows us to run.