Fixed initialization error in fame.
I was forgetting to zero the Mcoll matrix at the start of each call to
network_applyApproximateMethod, where it is created each time. gfortran seems
to like to reuse blocks of memory without rezeroing, so I was getting some
corrupt memory. This is why the RS method was failing so miserably.