Kinetics classes have new fixBarrier(p_Hrxn) method to fix activation energy.
In the case of ArrheniusEPKinetics (Evans-Polanyi) this evaluates the actual
barrier height and returns a regular ArrheniusKinetics instance.

In both cases (ArrheniusEPKinetics and ArrheniusKinetics) various checks are made
that the barrier is appropriate. If any changes are made to it, they are detailed
in the comments string.

The checks are:
 * Barrier isn't negative due to Evans-Polanyi estimate
   (intentional negative barriers are allowed)
 * Barrier is at least equal to the endothermicity for endothermic reactions.