Add cyclic-C3O and cyclopropyne to ForbiddenStructures.txt
Cyclic C3O is a saddle point and immediately becomes 
linear C3O according to http://dx.doi.org/10.1039/b718817j

Cyclopropyne is unstable according to http://dx.doi.org/10.1021/ja960762n
but note that cyclopropene-1,2-diyl, which is the same geometry but with a
double bond and 2 radicals instead of a triple bond, is stable and now in
the primaryThermoLibrary.

These were both causing problems with the automatic quantum calculations
of cyclic species.