Updating how RMG checks if node is an actual child of another node
When reading the frequency_groups database, RMG would write lines to the
output stating "____ is not actually a child of ___".  Thus, we thought
the tree was malformed.

However, the source of the problem was RMG's internal methods of determining
if a node was a child of another node.  If RMG was attempting to check if a
FunctionalGroupCollection was a child of a FunctionalGroup, it would
automatically return false, no questions asked.  This is the source of all
of the "___ is not actually a child of ___" for the frequency database.

This commit adds additional code to the FunctionalGroupCollection class,
which allows for checking whether a FunctionalGroupCollection is a child of
a FunctionalGroup.

As this resolves all gas-phase databases regarding the "___ is not actually
a child of ___" issue, MRH is closing the issue.

Closes #30