Fix enlargeReactionModel.
We create two lists, coreUpdateList and leakUpdateList which should each be of 
length rxnSystemList.size(). Blank entries should be `null` (as opposed to non-existant).
In the case of nothing to add to either list, we were forgetting to add `null`s 
so the lists were too short. 

Hopefully this closes #132