Use getFullName() in a few more places where it would help the output.
Sometimes we want to report the full name and species ID, but previously
to do so was a pain. Now it's easier, we can do it more often.

Hopefully I've not caught any places in which we don't actually know the ID,
or missed too many places where this would help. I was a little nervous around the sensitivity analysis section.