Cleaning up printing of memory use statistics.
Before we were using Runtime.totalMemory() as if it was the amount of memory
used. In actuality, the memory used is given by totalMemory() - freeMemory().
This explains why we haven't understood the memory output in the past: it
wasn't printing the right value!

We don't need to print memory usage every time we garbage collect; only once
per iteration is best. Every print should also use printMemoryUsage() for
consistency.