Check that limiting reactant name exists before looking for ID
After reading in the name of the limiting reactant, RMG would attempt to
find the species ID, using the species as the key.  However, if the species did
not exist (there was an error in the condition.txt file), RMG would throw
some exception within an unclear message.

Now, RMG checks that the species exists before checking for the species ID.
If the species does not exist, a clearer error message is given to the user.