Print execution statistics and model size information together in new log.
Before we printed this information after solving the reaction systems, when it
makes more sense to print it after each model enlargement step, since the
model size does not change during the reaction system solves. It is also
useful to print the memory use and running time next to the model size, as we
did in the original log file.