Reduce range of NASA thermo polynomials created by GATPFit.
The bigger range causes problems when Chemkin is pre-processing `tran.dat` transport file. I finally figured out it's this change to the thermo data that causes it, not anything in the `tran.dat` file. I'm reducing the range a bit to something that doesn't currently cause me problems.

See 886e01c90748ecec5e68959231e928749608271c where the range was expanded.

Closes #18 again.