BUG FIX: Third-body colliders spread across multiple lines
When reading the "pdepreactions.txt" file in any kinetics_libraries directory,
RMG assumed all third-body colliders would appear on one line.  If the third-
body colliders appeared on multiple lines (as is true for the Dooley\C1
mechanism), RMG would read-in the first line of third-body colliders properly
and save the instances.  However, upon finding additional third-body colliders,
RMG would then read these in properly and overwrite the previously-read
colliders.

This commit resolves this issue.  Rather than define the thirdBodyList
HashMap when reading in a colliders list, the read-in colliders are added to
the HashMap.

Closes #144