Terminate RMG when encountering an error in reading pdepreactions.txt file
Before: If RMG had any problem reading in the pdepreactions.txt file, it would
print an error message (that would quickly get lost in the sea of other
error/warning messages), store no pressure-dependent reactions in the Seed
Mechanism, and carry on.

Now: The error message printed to screen is a bit more detailed, and the RMG
simulation will terminate immediately.

Closes #145