Updates to documentation: conditionfile.txt
Majority of changes were specifying, in bold, if each field is Required or
Optional.

One other addition was a note about Chebyshev polynomials, namely that a user
should consider the max/min values of temperatures/pressures used in
calculating k(T,P) so that CHEMKIN (or similar) simulations behave properly.