Fixed bug in how edge reactions are counted in printModelSize().
I had assumed that the edge reactions were stored in the same way as the
core reactions, but this is not the case. The core reaction list
contains both the forward and reverse reactions separately; these must
be counted as 1/2 to avoid double-counting. However, the edge reaction
list only contains the forward direction, so these should be counted as
1.

The effect of this commit is to cause the number of edge reactions in
non-pressure-dependence jobs to double when printing the model size. The
generated model is otherwise unaffected.