Use McGowan's V (Abraham parameter), not UNIFAC volume, for solvent radius.
The McGowan's Volume (the V parameter) is pretty rudimentary but therefore robust.
As this is used to estimate diffusivities via Stokes-Einstein, we do not know
if one radius definition works better than another. Thus, we use the simplest.
This should give consistency, and means we don't need to read in the (malformed)
UNIFAC tree and dictionary.