Added Logger.logStackTrace() method.
You should use this method over Throwable.printStackTrace() to ensure
that the stack trace is printed both to the console and to the log file.
I have designed it to print the stack trace in a similar style as the
printStackTrace() method (at least for the version of Java I am using).

All printStackTrace() calls have been replaced with
Logger.logStackTrace() calls except for the runnable modules that are
not RMG.