Hiding NPE relating to Issue #130
This commit ensures that the NPE mentioned in Issue #130 will not occur.
However, this commit does NOT address why the NPE occurred in the first place.
This issue should be addressed in the near future; however, to promote a
working RMG 3.3, the NPE will be hidden for the time being.