Returning better "source" message in PopulateReactions class output
In the species thermochemistry section, the name, thermochemical data, and
source of each species are reported.  Currently, the "comments" are
returned, which usually results in many "null" printings.

Instead of returning the "comments", MRH is electing to return the "source".
The source already has detailed information about any PrimaryThermoLibrary.
In the event the source is null, MRH instead returns "Estimated by Group
Additivity".