Updated RMG example batch scripts to not redirect to RMG.log.
We don't need to do this anymore because RMG automatically saves the log
file. This is nice because the end user can now see the RMG output in
the console instead of just "Running RMG..." until the job is done. In
fact, without this change RMG will fail because it won't be able to
write to RMG.log from within Java when it's already open for handling
the redirected text.

Note that the other sets of examples should stay as-is because they do
not use the new Logger framework.