Flush the contents of the Final_Model.txt file
The Final_Model.txt file was never closed, so an empty file would be written
for each RMG run.  This commit resolves the issue.

The documentation (analyzing.txt) has also been updated (MRH was under the
impression only the first Reaction System's details were reported in the
Final_Model.txt file, but that was a bad assumption).