Removing potential NPE
In the (obsolete?) readIncludeSpecies function, the next read-in line is
sent to the Logger.info() function.  However, there is no check that the next
read-in line is not null.

MRH is removing this potential NPE.  Not being familiar with all of the new
RMG logging, I am not sure where to put (if we should put) this line, in this
particular function.

In fact, this entire readIncludedSpecies() function should probably be
removed entirely.