Removing the "IncludedSpecies" option from RMG
The "IncludedSpecies" option was put into RMG to allow low-flux species to be
added to the core, despite having
	a) a flux less than other edge species AND
	b) a flux less than the characteristic flux
The point was to add special species to the core immediately upon making it
via any reaction pathway.  (This may date back to 2007 ... MRH is not 100%
certain).

MRH believes this can now be done by adding these special species to a Seed
Mechanism.  While this is not mechanistically the same, it accomplishes
the same feat.