Seed Mechanisms reactions now added to core or edge properly.
Previously, they were all placed on the edge. This was a problem when a reaction
should have been in the core (based on all reagents being in the core in
the condition file). Such reactions are now placed in the core.

Also, we now add both forward and reverse directions, if reverse directions exist.

The addReactionSet(LinkedHashSet) method did almost exactly what I needed, but
it would discard reactions that had a reactant not in the core (you couldn't
have something entirely in the edge). Rather than change the behaviour
of that method (which is used elsewhere) I made a new method: addReactionSetFromSeed