Checking for duplicate species in user-specified input files
After RMG makes any species, it checks against its list of already-made species
to ensure it is not adding a duplicate.

When reading in a user-specified input file for any class within the RMG
package, the check still occurs, however if the check determines a
user-specified species is a duplicate of another user-specified species, we
assume that is an error on the user's part and immediately terminate the
simulation.  An error message, including the names of the offending species,
is supplied to the user.

Closes #152