Commenting 100 atm lines in Glarborg/C2 and C3 libraries
When the Glarborg reaction families were constructed, it appears the author
decided to use 10 atm as the "pressure of interest".  This is important because
several Glarborg reactions' kinetics are quoted at multiple pressures (or a
general correlation is given).  While all of the kinetics from the Glarborg
mechanism are listed in the RMG files, only the 10-atm values are not comments.

Having said that, MRH found 3 instances in the C2 and C3 libraries where both
the 10-atm and 100-atm values were not comments.  RMG interprets these as
duplicate kinetics and thus uses BOTH sets of kinetics when constructing a
reaction mechanism.

This commit resolves the issue (I have made the 100-atm kinetics comments, to
be consistent with the rest of the library).