Removing duplicate code re: LibraryReactionGenerator
In the ReactionModelGenerator class:
RMG reacts all initial core species against the
LibraryReactionGenerator (i.e. Reaction Library templates) to initialize
a LinkedHashSet.  The next step is to add these reactions (assuming they exist)
to another LinkedHashSet.

However, rather than simply addAll the just-generated LinkedHashSet to the
already-present LinkedHashSet, we re-generate the entire list.  This seems
silly to MRH, so I have moved the addAll call within the else (not empty)
statement and use the just-generated LinkedHashSet as the input argument.