Checking against "pdep" reactions in ReactionLibrary reactions.txt file
This commit switches the general order of adding reactions to the ODE input
file and CHEMKIN chem.inp file from:
   * Seed
   * Reaction Library pdepreactions
   * RMG (fame) pdepreactions
   * RMG nonpdep reactions

to:
   * Seed
   * Reaction Library pdepreactions
   * RMG nonpdep reactions
   # RMG (fame) pdepreactions

Example: Suppose user started with H and O2 in the input file, and Glarborg/C3
reaction library.  RMG would find H+O2=O+OH in Glarborg/C3 library and would
make HO2 (H+O2) included, thereby finding the O+OH well, making
H+O2(+m)=O+OH(+m).  In the previous order, RMG would add the RMG pdepreactions
reaction to the list, and then add the Glarborg/C3 reaction to the list: this
would be an incorrect duplicate.

With the new order, RMG adds the Glarborg/C3 reaction, and then recognizes
the RMG (fame) pdepreaction to be a duplicate and will not add it to the ODE
or CHEMKIN list.