Fix PDepReaction.setHighPKinetics(Kinetics kin)
The previous function would lead to a
null pointer exception if kinetics was null, and to an index out of bounds
exception if kinetics was a Kinetics[] array, because it would try to set
the kinetics[-1] element. I'm not sure what it was trying to achieve,
because I don't think it was ever used.

This creates a new empty kinetics array of size 1, and put the Kinetics in it.