Read in Chebyshev and PLog reactions from Seed Mechanisms and Reaction Libraries
The Reaction Library code re-uses the Seed Mechanism code for reading libraries,
so this affects both.

The kinetics on the line with the reaction definition are assumed to be the
high pressure limit kinetics (although in RMG-generated chemkin files with Chebyshev
format these currently read "1.0  0.0  0.0" so beware using this for reading those!
and perhaps restart files?)